include cell_line_token and adjust tables SAMPLES, ALLELES and FILES (#64)

+++ src/patches.py
@@ -9,6 +9,7 @@ implements retroactive changes to the database or config files
 '''
 
 import os, sys
+import re
 from configparser import ConfigParser
 
 from PyQt5.QtWidgets import (QApplication, QLabel, QMessageBox,
@@ -62,10 +63,18 @@ class GetPatchInput(QDialog):
         self.ipd_nr_fld.setText("0")
         layout.addRow(QLabel("IPD Submission Nr:"), self.ipd_nr_fld)
         
-        self.ok_btn = ProceedButton("Done", [self.ipd_short_fld, self.ipd_nr_fld], self.log)
+        intro_lbl3 = QLabel("(3) A short identifier for your company to use for cell line naming. An acronym etc. is ideal:")
+        intro_lbl3.setStyleSheet(general.label_style_italic)
+        layout.addRow(intro_lbl3)
+        self.token_fld = QLineEdit(self)
+        self.token_fld.setPlaceholderText("Can be identical to IPD shortname; max. 10 characters, only alphanumeric or '-'.")
+        layout.addRow(QLabel("Cell line identifier:"), self.token_fld)
+        
+        self.ok_btn = ProceedButton("Done", [self.ipd_short_fld, self.ipd_nr_fld, self.token_fld], self.log)
         layout.addRow(self.ok_btn)
         self.ipd_short_fld.textChanged.connect(self.ok_btn.check_ready)
         self.ipd_nr_fld.textChanged.connect(self.ok_btn.check_ready)
+        self.token_fld.textChanged.connect(self.ok_btn.check_ready)
         self.ok_btn.clicked.connect(self.on_clicked)
             
     def on_clicked(self):
@@ -73,19 +82,36 @@ class GetPatchInput(QDialog):
         """
         ipd_short = self.ipd_short_fld.text().strip()
         ipd_nr = self.ipd_nr_fld.text().strip()
+        token = self.token_fld.text().strip()
         self.log.debug("\t=> IPD Shortname: {}".format(ipd_short))
         self.log.debug("\t=> IPD Submission Nr: {}".format(ipd_nr))
+        self.log.debug("\t=> Cell line identifier: {}".format(token))
         self.log.info("Writing adjusted config files...")
+        
+        #sanity checks:
         try:
             int(ipd_nr)
         except ValueError:
             QMessageBox.warning(self, "This is not a number!", "Please specify a number in the field 'IPD Submission Nr'!")
+            self.ok_btn.setChecked(False)
+            return
+        pattern = "[^a-zA-Z0-9\-]+"
+        if re.search(pattern, token):
+            QMessageBox.warning(self, "Identifier rejected", "Please use only letters, numbers or '-' in the Cell line identifier!")
+            self.ok_btn.setChecked(False)
+            return
+        if len(token) > 10:
+            self.ok_btn.setChecked(False)
+            QMessageBox.warning(self, "Identifier rejected", "Your cell line identifier is too long.\nPlease restrict yourself to max. 10 characters!")
+            return
         
+        # write config files:
         self.log.debug("\tWriting patch config file...")
         with open(patch_config_file, "w") as g:
             g.write("[Company]\n")
             g.write("ipd_shortname = {}\n".format(ipd_short))
             g.write("ipd_submission_length = 7\n")
+            g.write("cell_line_token = {}\n".format(token))
             g.write("last_tl_version = 2.1.0\n")
             
         self.log.debug("\tWriting submission counter file...")
@@ -104,7 +130,8 @@ def check_patching_necessary(log):
     """checks config files for missing values
     """
     log.debug("Checking if any config patches necessary...")
-    patchme_dic = {company_config_file : {"Company" : ["ipd_shortname", "ipd_submission_length", "last_tl_version"]}
+    patchme_dic = {company_config_file : {"Company" : ["ipd_shortname", "ipd_submission_length", 
+                                                       "last_tl_version", "cell_line_token"]}
                    }
     needs_patching = False
     for config_file in patchme_dic:
@@ -198,25 +225,29 @@ pass
 #===========================================================
 # functions for database patches:
 
-def remove_cell_line_from_files(settings, conn, cursor, log):
-    log.info("Removing column cell_line from table FILES...")
+def replace_cell_line_in_FILES(conn, cursor, log):
+    """table FILES: replaces column cell_line with local_name
+    """
+    log.info("Replacing column cell_line of table FILES with local_name...")
     q0_drop_new_table = "DROP TABLE FILES_NEW" #if table remains from error during last attempt
     try:
         cursor.execute(q0_drop_new_table)
-    except: # if it's not there (it should not be there!)
+    except: # if it's not there that's fine (it should not be there!)
         pass
     
     q1_new_table = """CREATE TABLE FILES_NEW 
-                (SAMPLE_ID_INT TEXT , ALLELE_NR INT , LOCAL_NAME TEXT, PROJECT TEXT , 
+                (SAMPLE_ID_INT TEXT , ALLELE_NR INT , LOCAL_NAME TEXT PRIMARY KEY, PROJECT TEXT , 
                 RAW_FILE_TYPE TEXT , RAW_FILE TEXT , FASTA TEXT , BLAST_XML TEXT , 
                 ENA_FILE TEXT , ENA_RESPONSE_FILE TEXT , IPD_SUBMISSION_FILE TEXT )"""
     cursor.execute(q1_new_table)
-    q2_copy_data = """INSERT INTO FILES_NEW (SAMPLE_ID_INT, ALLELE_NR, LOCAL_NAME, PROJECT, 
-                RAW_FILE_TYPE, RAW_FILE, FASTA, BLAST_XML, 
+    q2_copy_data = """INSERT INTO FILES_NEW 
+                (SAMPLE_ID_INT, ALLELE_NR,
+                PROJECT, RAW_FILE_TYPE, RAW_FILE, FASTA, BLAST_XML, 
                 ENA_FILE, ENA_RESPONSE_FILE, IPD_SUBMISSION_FILE)
-               SELECT SAMPLE_ID_INT, ALLELE_NR, '', PROJECT, 
-                RAW_FILE_TYPE, RAW_FILE, FASTA, BLAST_XML, 
-                ENA_FILE, ENA_RESPONSE_FILE, IPD_SUBMISSION_FILE FROM FILES"""
+            SELECT SAMPLE_ID_INT, ALLELE_NR, 
+                PROJECT, RAW_FILE_TYPE, RAW_FILE, FASTA, BLAST_XML, 
+                ENA_FILE, ENA_RESPONSE_FILE, IPD_SUBMISSION_FILE 
+            FROM FILES"""
     cursor.execute(q2_copy_data)
     q3_drop_old_table = "DROP TABLE FILES"
     cursor.execute(q3_drop_old_table)
@@ -225,8 +256,9 @@ def remove_cell_line_from_files(settings, conn, cursor, log):
     conn.commit()
     log.info("\t=> Done")
 
-def add_missing_local_names_to_files_table(settings, conn, cursor, log):
-    """adds missing local names to table FILES
+
+def add_missing_local_names_to_FILES(conn, cursor, log):
+    """table FILES: adds missing local names
     """
     log.info("Adding local_names to table FILES...")
     log.info("\tGetting local_names from table ALLELES...")
@@ -237,9 +269,91 @@ def add_missing_local_names_to_files_table(settings, conn, cursor, log):
     update_query = "update or ignore FILES set local_name = :1 where sample_id_int = :2 and allele_nr = :3"
     cursor.executemany(update_query, items)
     conn.commit()
-    log.info("\t=> Done")    
+    log.info("\t=> Done")  
+
+
+def add_cell_line_to_SAMPLES(settings, conn, cursor, log):
+    """table SAMPLES: adds missing column cell_line
+    """
+    log.info("Adding column cell_line to table SAMPLES...")
+    q1_add_column = "alter table SAMPLES add column CELL_LINE"
+    cursor.execute(q1_add_column)
+    
+    log.debug("\tGetting existing samples...")
+    q2_get_samples = "select sample_id_int, sample_id_ext from SAMPLES"
+    items = db_internal.query_database(q2_get_samples, None, log, cursor)
+    
+    items = [('{}_{}'.format(settings["cell_line_token"], item[0]), item[0], item[1]) for item in items]
+    log.info("\tWriting local_names into table FILES...")
+    update_query = "update SAMPLES set cell_line = :1 where sample_id_int = :2 and sample_id_ext = :3"
+    cursor.executemany(update_query, items)
+    conn.commit()
+    log.info("\t=> Done") 
     
     
+def change_pk_of_ALLELES(conn, cursor, log):
+    """table ALLELES: moves primary key to column local_name
+    """
+    log.info("Changing primary key of table ALLELES to column local_name and renames column cell_line...")
+    q0_drop_new_table = "DROP TABLE ALLELES_NEW" #if table remains from error during last attempt
+    try:
+        cursor.execute(q0_drop_new_table)
+    except: # if it's not there that's fine (it should not be there!)
+        pass
+    
+    q1_new_table = """CREATE TABLE ALLELES_NEW 
+                (SAMPLE_ID_INT TEXT , ALLELE_NR INT , PROJECT_NAME TEXT , PROJECT_NR INT , 
+                LOCAL_NAME TEXT PRIMARY KEY, GENE TEXT , GOAL TEXT , ALLELE_STATUS TEXT , 
+                UPLOAD_DATE TEXT, DETECTION_DATE TEXT, ORIG_ALLELE1 TEXT , 
+                ORIG_ALLELE2 TEXT , ORIG_GENOTYPING_SOFTWARE TEXT , ORIG_SOFTWARE_VERSION TEXT , 
+                ORIG_GENOTYPING_DATE TEXT , LAB_STATUS TEXT , PANEL TEXT , POSITION TEXT , SHORT_READ_DATA TEXT , 
+                SHORT_READ_PHASING TEXT , SHORT_READ_TECHNOLOGY TEXT , LONG_READ_DATA TEXT , 
+                LONG_READ_PHASING TEXT , LONG_READ_TECHNOLOGY TEXT , KOMMENTAR TEXT , TARGET_ALLELE TEXT , 
+                PARTNER_ALLELE TEXT , MISMATCH_POSITION TEXT , NULL_ALLELE TEXT , 
+                NEW_GENOTYPING_SOFTWARE TEXT , NEW_SOFTWARE_VERSION TEXT , NEW_GENOTYPING_DATE TEXT , 
+                REFERENCE_DATABASE TEXT , DATABASE_VERSION TEXT , INTERNAL_NAME TEXT , OFFICIAL_NAME TEXT , 
+                NEW_CONFIRMED TEXT , ENA_SUBMISSION_ID TEXT , ENA_ACCEPTION_DATE TEXT , ENA_ACCESSION_NR TEXT , 
+                IPD_SUBMISSION_ID TEXT , IPD_SUBMISSION_NR TEXT , HWS_SUBMISSION_NR TEXT , IPD_ACCEPTION_DATE TEXT , 
+                IPD_RELEASE TEXT, CELL_LINE_OLD TEXT)
+                """
+    cursor.execute(q1_new_table)
+    q2_copy_data = """
+        INSERT INTO ALLELES_NEW 
+            (SAMPLE_ID_INT, ALLELE_NR, PROJECT_NAME, PROJECT_NR,
+            LOCAL_NAME, GENE, GOAL, ALLELE_STATUS, upload_date, detection_date, 
+            ORIG_ALLELE1, ORIG_ALLELE2, ORIG_GENOTYPING_SOFTWARE, ORIG_SOFTWARE_VERSION, 
+            ORIG_GENOTYPING_DATE, LAB_STATUS, PANEL, POSITION, SHORT_READ_DATA, 
+            SHORT_READ_PHASING, SHORT_READ_TECHNOLOGY, LONG_READ_DATA, 
+            LONG_READ_PHASING, LONG_READ_TECHNOLOGY, KOMMENTAR, TARGET_ALLELE, 
+            PARTNER_ALLELE, MISMATCH_POSITION, NULL_ALLELE, 
+            NEW_GENOTYPING_SOFTWARE, NEW_SOFTWARE_VERSION, NEW_GENOTYPING_DATE, 
+            REFERENCE_DATABASE, DATABASE_VERSION, INTERNAL_NAME, OFFICIAL_NAME, 
+            NEW_CONFIRMED, ENA_SUBMISSION_ID, ENA_ACCEPTION_DATE, ENA_ACCESSION_NR, 
+            IPD_SUBMISSION_ID, IPD_SUBMISSION_NR, HWS_SUBMISSION_NR, IPD_ACCEPTION_DATE, 
+            IPD_RELEASE, CELL_LINE_OLD)
+        SELECT 
+            SAMPLE_ID_INT, ALLELE_NR, PROJECT_NAME, PROJECT_NR, 
+            LOCAL_NAME, GENE, GOAL, ALLELE_STATUS, upload_date, detection_date,
+            ORIG_ALLELE1, ORIG_ALLELE2, ORIG_GENOTYPING_SOFTWARE, ORIG_SOFTWARE_VERSION, 
+            ORIG_GENOTYPING_DATE, LAB_STATUS, PANEL, POSITION, SHORT_READ_DATA, 
+            SHORT_READ_PHASING, SHORT_READ_TECHNOLOGY, LONG_READ_DATA, 
+            LONG_READ_PHASING, LONG_READ_TECHNOLOGY, KOMMENTAR, TARGET_ALLELE, 
+            PARTNER_ALLELE, MISMATCH_POSITION, NULL_ALLELE, 
+            NEW_GENOTYPING_SOFTWARE, NEW_SOFTWARE_VERSION, NEW_GENOTYPING_DATE, 
+            REFERENCE_DATABASE, DATABASE_VERSION, INTERNAL_NAME, OFFICIAL_NAME, 
+            NEW_CONFIRMED, ENA_SUBMISSION_ID, ENA_ACCEPTION_DATE, ENA_ACCESSION_NR, 
+            IPD_SUBMISSION_ID, IPD_SUBMISSION_NR, HWS_SUBMISSION_NR, IPD_ACCEPTION_DATE, 
+            IPD_RELEASE, CELL_LINE
+        FROM ALLELES"""
+    cursor.execute(q2_copy_data)
+    q3_drop_old_table = "DROP TABLE ALLELES"
+    cursor.execute(q3_drop_old_table)
+    q4_rename_new_table = "ALTER TABLE ALLELES_NEW RENAME TO ALLELES"
+    cursor.execute(q4_rename_new_table)
+    conn.commit()
+    log.info("\t=> Done")
+    
+
 def patch_database(settings, log):
     """patches the SQLite database of the current user
     """
@@ -253,8 +367,10 @@ def patch_database(settings, log):
     
     conn, cursor = db_internal.open_connection(settings["db_file"], log)
     
-    remove_cell_line_from_files(settings, conn, cursor, log)
-    add_missing_local_names_to_files_table(settings, conn, cursor, log)
+    add_cell_line_to_SAMPLES(settings, conn, cursor, log)
+    change_pk_of_ALLELES(conn, cursor, log)
+    replace_cell_line_in_FILES(conn, cursor, log)
+    add_missing_local_names_to_FILES(conn, cursor, log)
     
     cursor.close()
     conn.close()
